Feat: 메인+랭킹 페이지 lazy 설정 및 로딩 중 AppSpinner 적용
