+++ src/components/AppSpinner/AppSpinner.module.css
@@ -1,10 +1,10 @@
 .appSpinner {
   width: 100%;
-  height: 100%;
   display: flex;
   flex-flow: column;
   align-items: center;
+  justify-content: center;
   gap: var(--space-4, 16px);
   margin: 0 auto;
-  padding: var(--space-10, 40px) 0;
+  padding: 200px 0;
 }
